+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+- [Feature] Add `copyButtonStyle` prop to **Code** component, and `style` prop to **CopyButton** component ([#1152](https://github.com/optimizely/oui/pull/1152))
 - [Patch] Update positioning for the **Sheet** body content ([#1148](https://github.com/optimizely/oui/pull/1148))
 
 ## 42.7.0 - 2019-04-19